Improvement: Updating the networks and config files with custom anchors for CITYSCAPE dataset
This commit updates the config files and the associated networks to make use of custom anchors designed for the CITYSCAPE dataset.

Note: Known issue is that the ResNet16 network needs to be updated to make use of the latest anchors. The other networks have been tested. VGG16 works without pretrained weights, however the pre-trained weigths seem to be broken.